requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1671589829431/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1681349778161/work
certifi==2022.12.7
charset-normalizer==3.1.0
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
contourpy==1.0.7
cycler==0.11.0
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1680693679482/work
filelock==3.12.0
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1682709117081/work
Flask-WTF==1.1.1
fonttools==4.39.3
gunicorn==20.1.0
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1648147185463/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
kiwisolver==1.4.4
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1674135859696/work
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.21.1
six==1.16.0
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1681817562700/work
sympy==1.11.1
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
ultralytics==8.0.90
urllib3==1.26.15
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1682779753483/work
WTForms==3.0.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
